+++ HISTORY.rst
@@ -19,6 +19,7 @@ Unreleased
 
 * ALCN-1118 : Fix Customer Address export when esb_ref is empty
 * ALCN-1106: Fix display alignment columns on sale order line detail form.
+ * ALCN-1121 Fix OnlinePayment and False in string values on customer export
 
 **Build**